Fix displayed MAC address
When in config mode, the OBI Steckdose works in AP mode providing a SSID based
on their MAC address. This way, it's easier to add it to some DHCP server.
However, the most significant byte of the AP MAC address is two higher than the
MAC address in STA mode. So fix this displayed MAC address.